adding cohereChat
