+++ app/src/App.js
@@ -1,15 +1,11 @@
-import logo from './logo.svg';
 import './App.css';
-import ChatRoom from './views/cohereController.js';
+import ChatRoom from './views/cohereChat.js';
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
+        
         <ChatRoom />
       </header>
     </div>
